agent/chat_bot/requirements.txt
fastapi
google-api-python-client
pydantic
python-dotenv
uvicorn[standard]
wikipedia
httpx
google-generativeai
